[FIX] 干掉Parser拷贝内存接口，不拷贝内存 [Feat] 增加render接口

+++ VideoCodecKitDemo/VideoCodecKitDemo/VideoCodecKit/BaseModel/VCH264Frame.h
@@ -31,8 +31,8 @@ typedef NS_ENUM(NSUInteger, VCH264FrameType) {
 @property (nonatomic, assign) NSUInteger height;
 @property (nonatomic, strong) VCVideoFPS *fps;
 
-@property (nonatomic, assign) NSInteger pts;
-@property (nonatomic, assign) NSInteger dts;
+@property (nonatomic, assign) int64_t pts;
+@property (nonatomic, assign) int64_t dts;
 
 @property (nonatomic, assign) uint8_t *parseData;
 @property (nonatomic, assign) NSUInteger parseSize;
